moved print_thanks to ui module and updated imports
